Added a LICENSE.md to specify an MIT license. Added a Footer component that is rendered on each page. Added a landing page that the user will be presented with upon logging in to their account.
